rename to b-divider.vue closes #263
