+++ lib/components/index.js
@@ -9,7 +9,7 @@ import bCarouselSlide from './carousel-slide.vue';
 import bCollapse from './collapse.vue';
 import bDropdown from './dropdown.vue';
 import bDropdownItem from './dropdown-item.vue';
-import bDropdownDivider from './dropdown-devider.vue';
+import bDropdownDivider from './dropdown-divider.vue';
 import bDropdownSelect from './dropdown-select.vue';
 import bForm from './form.vue';
 import bFormFieldset from './form-fieldset.vue';
